+++ src/sprites/__init__.py
@@ -1,2 +1,3 @@
 from .gem import Gem
 from .player import Player
+from .enemy import Enemy