+++ src/tests/game_engine/game_state_test.py
@@ -49,3 +49,23 @@ class GameStateTest(unittest.TestCase):
         self.game_state.add_points(300)
         self.game_state.add_points(-100)
         self.assertEqual(300, self.game_state.points)
+
+    def test_game_state_starts_with_no_enemies(self):
+        self.assertEqual(0, len(self.game_state.enemies))
+
+    @patch("sprites.enemy.image_handler.load_image")
+    def test_spawn_enemy_creates_an_enemy(self, mock_surface):
+        mock_surface.return_value = pygame.Surface((32, 40))
+
+        for i in range(2):
+            with self.subTest(enemy_count=i + 1):
+                self.game_state.spawn_enemy()
+                self.assertEqual(i + 1, len(self.game_state.enemies))
+
+    @patch("sprites.enemy.image_handler.load_image")
+    def test_spawn_enemy_creates_an_enemy_with_correct_speed_value(self, mock_surface):
+        mock_surface.return_value = pygame.Surface((32, 40))
+
+        self.game_state.spawn_enemy(50)
+        enemy = self.game_state.enemies.sprites()[0]
+        self.assertEqual(50, enemy.speed)